refactor: consume shared Git kernel from plugin-git-lib
Phase 1 of kestra-io/kestra-ee#7117: deletes the connection, authentication
and clone/push/sync plumbing that now lives in plugin-git-lib
(io.kestra.plugin.git.shared) and depends on it as
io.kestra.plugin:plugin-git-lib:1.0.0-SNAPSHOT.

- AbstractGitTask/AbstractKestraTask/AbstractCloningTask/AbstractSyncTask/AbstractPushTask,
  KestraApiAuth/KestraApiConnection and services/GitService,SshTransportConfigCallback moved
  out; concrete tasks (Clone, Push*, Sync*, NamespaceSync, TenantSync) stay here and import
  the shared classes.
- Clone now delegates its clone/checkout logic to the lib's CloneService instead of
  duplicating it.
- Drops the now-transitive jgit and kestra-api-client dependency declarations.
- Test scaffolding (AbstractGitTest, AbstractKestraContainerTest,
  KestraContainerTestDataUtils, MockKestraApiServer) now consumed from the lib's
  testFixtures; KestraApiConnectionTest/AbstractKestraTaskAuthTest removed here since
  equivalent coverage now lives in the lib.
- Same 66 pre-existing network/credential-gated test failures as the unmodified baseline
  (no GitHub PAT / Gitea container available in this environment); no new failures.
