+++ src/main/java/io/kestra/plugin/git/NamespaceSync.java
@@ -40,7 +40,8 @@ import io.kestra.core.models.tasks.RunnableTask;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.serializers.YamlParser;
 import io.kestra.core.storages.NamespaceFile;
-import io.kestra.plugin.git.services.GitService;
+import io.kestra.plugin.git.shared.AbstractCloningTask;
+import io.kestra.plugin.git.shared.services.GitService;
 import io.kestra.sdk.internal.ApiException;
 import io.kestra.sdk.model.QueryFilter;
 import io.kestra.sdk.model.QueryFilterField;
