+++ src/main/java/io/kestra/plugin/git/Push.java
@@ -1,5 +1,7 @@
 package io.kestra.plugin.git;
 
+import io.kestra.plugin.git.shared.AbstractCloningTask;
+
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
