+++ src/main/java/io/kestra/plugin/git/PushExecutionFiles.java
@@ -1,5 +1,7 @@
 package io.kestra.plugin.git;
 
+import io.kestra.plugin.git.shared.AbstractPushTask;
+
 import java.io.FileInputStream;
 import java.io.InputStream;
 import java.net.URI;
