+++ src/main/java/io/kestra/plugin/git/SyncFlow.java
@@ -16,7 +16,8 @@ import io.kestra.core.models.annotations.PluginProperty;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.models.tasks.RunnableTask;
 import io.kestra.core.runners.RunContext;
-import io.kestra.plugin.git.services.GitService;
+import io.kestra.plugin.git.shared.AbstractKestraTask;
+import io.kestra.plugin.git.shared.services.GitService;
 import io.kestra.sdk.KestraClient;
 import io.kestra.sdk.internal.ApiException;
 import io.kestra.sdk.model.FlowWithSource;
