+++ src/main/java/io/kestra/plugin/git/TenantSync.java
@@ -31,7 +31,8 @@ import io.kestra.core.models.property.Property;
 import io.kestra.core.models.tasks.RunnableTask;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.serializers.YamlParser;
-import io.kestra.plugin.git.services.GitService;
+import io.kestra.plugin.git.shared.AbstractKestraTask;
+import io.kestra.plugin.git.shared.services.GitService;
 import io.kestra.sdk.KestraClient;
 import io.kestra.sdk.api.FilesApi;
 import io.kestra.sdk.internal.ApiException;
