--- src/main/java/io/kestra/plugin/git/services/SshTransportConfigCallback.java
@@ -1,52 +1,0 @@
-package io.kestra.plugin.git.services;
-
-import java.nio.charset.StandardCharsets;
-
-import org.eclipse.jgit.api.TransportConfigCallback;
-import org.eclipse.jgit.transport.SshTransport;
-import org.eclipse.jgit.transport.Transport;
-import org.eclipse.jgit.transport.ssh.jsch.JschConfigSessionFactory;
-import org.eclipse.jgit.transport.ssh.jsch.OpenSshConfig;
-import org.eclipse.jgit.util.FS;
-
-import com.jcraft.jsch.JSch;
-import com.jcraft.jsch.JSchException;
-import com.jcraft.jsch.Session;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class SshTransportConfigCallback implements TransportConfigCallback {
-    private byte[] privateKey;
-    private String passphrase;
-
-    @Override
-    public void configure(Transport transport) {
-        SshTransport sshTransport = (SshTransport) transport;
-        sshTransport.setSshSessionFactory(new JschConfigSessionFactory() {
-            @Override
-            protected void configure(OpenSshConfig.Host hc, Session session) {
-                session.setConfig("StrictHostKeyChecking", "no");
-            }
-
-            @Override
-            protected JSch getJSch(final OpenSshConfig.Host hc, FS fs) throws JSchException {
-                JSch jsch = super.getJSch(hc, fs);
-                jsch.removeAllIdentity();
-
-                if (passphrase != null) {
-                    jsch.addIdentity(
-                        "privateKey",
-                        privateKey,
-                        null,
-                        passphrase.getBytes(StandardCharsets.UTF_8)
-                    );
-                } else {
-                    jsch.addIdentity("privateKey", privateKey, null, null);
-                }
-
-                return jsch;
-            }
-        });
-    }
-}
