+++ src/test/java/io/kestra/plugin/git/SyncFlowsTest.java
@@ -1,5 +1,9 @@
 package io.kestra.plugin.git;
 
+import io.kestra.plugin.git.shared.AbstractSyncTask;
+import io.kestra.plugin.git.shared.testkit.AbstractGitTest;
+import io.kestra.plugin.git.shared.testkit.MockKestraApiServer;
+
 import java.io.*;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
